Rename API resource Version related classes to ResourceVersion name pattern
