+++ src/Filter/ResourceVersionResourceFilter.php
@@ -5,11 +5,25 @@ namespace App\Filter;
 use ApiPlatform\Doctrine\Orm\Filter\AbstractFilter;
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
 use ApiPlatform\Metadata\Operation;
+use App\Service\VersionedResourceService;
 use Doctrine\ORM\QueryBuilder;
+use Doctrine\Persistence\ManagerRegistry;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\PropertyInfo\Type;
+use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 
-final class VersionResourceIdFilter extends AbstractFilter
+final class ResourceVersionResourceFilter extends AbstractFilter
 {
+    public function __construct(
+        private VersionedResourceService $versionedResourceService,
+        protected ManagerRegistry $managerRegistry,
+        ?LoggerInterface $logger = null,
+        protected ?array $properties = null,
+        protected ?NameConverterInterface $nameConverter = null,
+    ) {
+        parent::__construct($managerRegistry, $logger, $properties, $nameConverter);
+    }
+
     protected function filterProperty(
         string $property,
         $value,
@@ -28,9 +42,13 @@ final class VersionResourceIdFilter extends AbstractFilter
         foreach ($this->properties as $property => $strategy) {
             $description["$property"] = [
                 'property' => $property,
-                'type' => Type::BUILTIN_TYPE_INT,
+                'type' => Type::BUILTIN_TYPE_STRING,
                 'required' => true,
-                'description' => 'The ID of the named resource.',
+                'description' => 'The name of the resource.',
+                'schema' => [
+                    'type' => Type::BUILTIN_TYPE_STRING,
+                    'enum' => $this->versionedResourceService->getNames()
+                ],
                 'openapi' => [
                     'allowEmptyValue' => false,
                 ],
